std::uint32_t --> std::int64_t

+++ code/BitsCount.h
@@ -1,6 +1,6 @@
 #include <cstdint>
 
-std::uint8_t bruteForce(std::uint32_t num);
-std::uint8_t algoKernighan(std::uint32_t num);
-std::uint8_t lookupBytes(std::uint32_t num);
-std::uint8_t divideAndConquer(std::uint32_t num);
+std::uint8_t bruteForce(std::int64_t num);
+std::uint8_t algoKernighan(std::int64_t num);
+std::uint8_t lookupBytes(std::int64_t num);
+std::uint8_t divideAndConquer(std::int64_t num);
